pyproject.toml
[tool.poetry]
name = "ragagent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
a2a-sdk = "^0.2.10"
uvicorn = "^0.35.0"
click = "^8.2.1"
httpx = "^0.28.1"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
langchain-google-genai = "^2.1.6"
langgraph = "^0.5.0"
pytest = "^8.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
